Rework tests using pytest
Also cleaned up test-requirements.txt
