+++ test-requirements.txt
@@ -1,19 +1,2 @@
-hacking<0.10,>=0.9.2
-
-bandit>=1.1.0 # Apache-2.0
-coverage>=3.6
-discover
-fixtures>=1.3.1
-python-subunit>=0.0.18
-sphinx!=1.2.0,!=1.3b1,<1.3,>=1.1.2
-# oslosphinx>=2.5.0 # Apache-2.0
-sphinx-argparse
-testrepository>=0.0.18
-testscenarios>=0.4
-testtools>=1.4.0
-
-mock>=1.2
-nose>=1.1.2
-nosexcover
-
 regex
+pytest
